Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -7,18 +7,16 @@
  */
 int main(void)
 {
-	char c;
-	int i;
-	long int l;
-	long long int ll;
-	float d;
+char c;
+int i;
+long int l;
+long long int ll;
+float d;
 
-printf("Size of a char: %lu byte(s)\n", sizeof(c));
-printf("Size of a int: %lu byte(s)\n", sizeof(i));
-printf("Size of a long int: %lu byte(s)\n", sizeof(l));
-printf("Size of a long long int: %lu byte(s)\n", sizeof(ll));
-printf("Size of a float: %lu byte(s)\n", sizeof(d));
-
+printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
+printf("Size of a int: %lu byte(s)\n", (unsigned long)sizeof(i));
+printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(l));
+printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(ll));
+printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(d));
 return (0);
-
 }
